astora_lights_if/requirements.txt
websocket-client==1.7.0
numpy==1.21.5